pyproject.toml
[project]
name = "greengrid-forecast"
version = "0.1.0"
description = "Energy consumption and RES production forecasting system."
authors = [
    {name = "Jakub Milczarczyk",email = "milczarczykjakub@gmail.com"}
]
license = {text = "Proprietary"}
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "polars (>=1.28.1,<2.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "openmeteo-requests (>=1.4.0,<2.0.0)",
    "entsoe-py (>=0.6.18,<0.7.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "streamlit (>=1.45.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
